Extract menu building to utility class
