+++ recaf-ui/src/main/java/me/coley/recaf/ui/context/ClassContextBuilder.java
@@ -25,6 +25,8 @@ import org.objectweb.asm.ClassWriter;
 
 import java.util.Optional;
 
+import static me.coley.recaf.ui.util.Menus.*;
+
 /**
  * Context menu builder for classes.
  *
