+++ recaf-ui/src/main/java/me/coley/recaf/ui/context/DexClassContextBuilder.java
@@ -9,6 +9,8 @@ import me.coley.recaf.util.StringUtil;
 import me.coley.recaf.workspace.Workspace;
 import me.coley.recaf.workspace.resource.Resource;
 
+import static me.coley.recaf.ui.util.Menus.*;
+
 /**
  * Context menu builder for android dex classes.
  *
